+++ src/utils/display_area.rs
@@ -1,3 +1,5 @@
+//! Read display area as a rectangle.
+
 use crate::prelude::*;
 use embedded_graphics::{primitives::Rectangle, DrawTarget};
 
@@ -8,7 +10,7 @@ where
 {
     /// Return the display area as a `Rectangle`
     ///
-    /// This method is provided mainly to make it simpler to align to the edges (or the center) of the display.
+    /// This method is provided mainly to make it simpler to align to the display.
     fn display_area(&self) -> Rectangle;
 }