+++ src/utils/mod.rs
@@ -1,7 +1,4 @@
 //! Utility collection module
 
-/// Helper to retrieve display area
 pub mod display_area;
-
-/// Rectangle extensions
 pub mod rect_helper;